+++ app/packages/looker/src/lookers/video.ts
@@ -219,6 +219,20 @@ export class VideoLooker extends AbstractLooker<VideoState, VideoSample> {
 
   get waiting() {
     const video = this.lookerElement.children[0].element as HTMLVideoElement;
+    if (
+      !this.state.config.thumbnail &&
+      video &&
+      (video.seeking ||
+        video.readyState < 2 ||
+        !this.hasFrame(this.state.frameNumber))
+    ) {
+      console.log(
+        video,
+        video.seeking,
+        video.readyState < 2,
+        !this.hasFrame(this.state.frameNumber)
+      );
+    }
     return (
       video &&
       (video.seeking ||
